Allow alternate Select option typeahead strings Previously, select options' text must always match their value for typeahead to work as expected.
    <option value="A">A</option>
    <option value="B">B</option>
    <option value="C">C</option>

This change allows the value and typeahead text to optionally differ.

    <option value="_1" typeahead="A">A</option>
    <option value="_2" typeahead="B">B</option>
    <option value="C">C</option>

This is useful for multilingual scenarios where the value may be a known
constant, but the text rendered on the option may vary from language
to language. Allowing alternate typeahead values here prevents having to
convert back-and-forth between unlocalized IDs and localized text.

`innerText` was considered, but avoided, since that property may trigger
reflows, where changing a dataset entry does not. Further, using a separate
value allows typeahead to function even when there is no text element at
all (for example, if the "text" is a stylized SVG).