Fix: Allow options in Select and Combobox to not be DOM-sequential (#97)
* Allow options in Select to not be DOM-sequential
Previously, options must be sequential in the DOM for keyboard navigation
to function as expected.

    <option />
    <option />
    <option />

This change allows options to not be sequential.

    <div>
        <option />
        <option />
    </div>
    <div>
        <option />
    </div>

This facilitates grouping select options or inserting other
presentational elements within a Select menu.

* Also use non-sequential keyboard navigation logic in Combobox

* Remove unnecessary optional access operator
The parameter to `[].find()` will only be called if the list has at least
one element. This means that `[].at()` will always return a valid entry,
since the range we're supplying -- `[-1, N)` -- will always be a valid
index.

* Create wild-tigers-nail.md

* readability

* yay

---------

Co-authored-by: Thomas G. Lopes <26071571+TGlide@users.noreply.github.com>